Luajit compat + initial `broom` impl
Also the beginnings of "lawful.fnl", which should eventually
become a shim around all of the `awful` and perhaps other default
awesome libraries.

Broom is meant as a rofi-like input prompt for doing fuzzy
selection. So far, just have support for a basic launcher but
have a levenshtein distance impl ready.